fix: 统一时间戳类型为i64并更新依赖版本
- 将trade模块中的timestamp字段从u64改为i64，避免类型转换错误
- 更新多个Cargo.toml中的依赖版本，包括async-trait、futures、uuid等
- 重构ImgGenNode使用统一的GenImgModel接口，移除冗余的HTTP请求代码
- 将Controller的RunnerId从u64改为u16以节省内存

+++ nodes/src/trade/source/tushare.rs
@@ -93,7 +93,7 @@ impl TushareSource {
                     .to_owned(),
                 timestamp: date_str_to_timestamp(
                     v["trade_date"].as_str().expect("trade_date field error"),
-                ) as u64,
+                ),
                 open: v["open"].as_f64().expect("open field error"),
                 high: v["high"].as_f64().expect("high field error"),
                 low: v["low"].as_f64().expect("low field error"),
